Organized models in domain sub-modules
