+++ domain-model/gps-tracker-model/src/main/java/com/ridesharing/gpstrackerservice/DeviceType.java
@@ -1,4 +1,4 @@
-package com.ridesharing.domain.model.ride.gps.position;
+package com.ridesharing.gpstrackerservice;
 
 public enum DeviceType {
     CLIENT_DEVICE,
