+++ domain-model/pom.xml
@@ -1,6 +1,8 @@
-<?xml version="1.0" encoding="UTF-8"?>
+<?xml version="1.0" encoding="UTF-8" standalone="no"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+
   <modelVersion>4.0.0</modelVersion>
+
   <parent>
     <artifactId>ride-sharing-root</artifactId>
     <groupId>com.ridesharing</groupId>
@@ -14,4 +16,11 @@
     <spring-boot.repackage.skip>true</spring-boot.repackage.skip>
   </properties>
 
+  <packaging>pom</packaging>
+
+  <modules>
+    <module>ride-planner-model</module>
+    <module>gps-tracker-model</module>
+  </modules>
+  
 </project>
