+++ domain-model/ride-planner-model/src/main/java/com/ridesharing/rideplannerservice/RideUpdateType.java
@@ -1,4 +1,4 @@
-package com.ridesharing.domain.model.ride;
+package com.ridesharing.rideplannerservice;
 
 public enum RideUpdateType {
     CLIENT_CANCELLATION,
