+++ domain-model/ride-planner-model/src/main/java/com/ridesharing/rideplannerservice/events/EventType.java
@@ -1,4 +1,4 @@
-package com.ridesharing.domain.model.ride.events;
+package com.ridesharing.rideplannerservice.events;
 
 public enum EventType {
     RIDE_CREATED,
