+++ gps-position-tracker-service/pom.xml
@@ -14,7 +14,7 @@
     <!-- Domain model -->
     <dependency>
       <groupId>com.ridesharing</groupId>
-      <artifactId>domain-model</artifactId>
+      <artifactId>gps-tracker-model</artifactId>
       <version>${project.version}</version>
       <scope>compile</scope>
     </dependency>
