+++ gps-position-tracker-service/src/main/java/com/ridesharing/gps/position/tracker/eventsender/EventSender.java
@@ -1,6 +1,7 @@
 package com.ridesharing.gps.position.tracker.eventsender;
 
-import com.ridesharing.domain.model.ride.gps.position.GpsPosition;
+
+import com.ridesharing.gpstrackerservice.GpsPosition;
 
 public interface EventSender {
     void send(GpsPosition newGpsPosition);
