+++ gps-position-tracker-service/src/main/java/com/ridesharing/gps/position/tracker/service/GpsPositionTrackerService.java
@@ -1,6 +1,6 @@
 package com.ridesharing.gps.position.tracker.service;
 
-import com.ridesharing.domain.model.ride.gps.position.GpsPosition;
+import com.ridesharing.gpstrackerservice.GpsPosition;
 
 public interface GpsPositionTrackerService {
     void notifyGpsPositionChanged(GpsPosition newGpsPosition);
