+++ gps-position-tracker-service/src/main/java/com/ridesharing/gps/position/tracker/service/GpsPositionTrackerServiceImpl.java
@@ -1,7 +1,7 @@
 package com.ridesharing.gps.position.tracker.service;
 
-import com.ridesharing.domain.model.ride.gps.position.GpsPosition;
 import com.ridesharing.gps.position.tracker.eventsender.EventSender;
+import com.ridesharing.gpstrackerservice.GpsPosition;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
