+++ gps-position-tracker-service/src/test/java/com/ridesharing/gps/position/tracker/service/controller/GpsPositionTrackerControllerTest.java
@@ -1,6 +1,6 @@
 package com.ridesharing.gps.position.tracker.service.controller;
 
-import com.ridesharing.domain.model.ride.gps.position.GpsPosition;
+import com.ridesharing.gpstrackerservice.GpsPosition;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -10,10 +10,9 @@ import org.springframework.web.reactive.function.BodyInserters;
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 class GpsPositionTrackerControllerTest {
 
-    private static final String BASE_URI = "/gps-receiver";
     public static final float LATITUDE = 38.897675f;
     public static final float LONGITUDE = -73.9729691f;
-
+    private static final String BASE_URI = "/gps-receiver";
     @Autowired
     private WebTestClient webClient;
