+++ gps-position-tracker-service/src/test/postman/GPS Tracker Service.postman_collection.json
@@ -1,40 +1,40 @@
 {
-	"info": {
-		"_postman_id": "a29bf7a0-71cf-484c-afc6-c5cc7da83301",
-		"name": "GPS Tracker Service",
-		"schema": "https://schema.getpostman.com/json/collection/v2.1.0/collection.json"
-	},
-	"item": [
-		{
-			"name": "Send new GPS coordinates to receiver",
-			"request": {
-				"method": "POST",
-				"header": [],
-				"body": {
-					"mode": "raw",
-					"raw": "{\r\n    \"trackedDeviceId\": \"ABBA\",\r\n    \"deviceType\": \"CLIENT_DEVICE\",\r\n    \"latitude\": 38.897675,\r\n    \"longitude\": -73.9729691,\r\n    \"speed\": 100\r\n}",
-					"options": {
-						"raw": {
-							"language": "json"
-						}
-					}
-				},
-				"url": {
-					"raw": "192.168.1.144:8082/gps-receiver/",
-					"host": [
-						"192",
-						"168",
-						"1",
-						"144"
-					],
-					"port": "8082",
-					"path": [
-						"gps-receiver",
-						""
-					]
-				}
-			},
-			"response": []
-		}
-	]
+  "info": {
+    "_postman_id": "a29bf7a0-71cf-484c-afc6-c5cc7da83301",
+    "name": "GPS Tracker Service",
+    "schema": "https://schema.getpostman.com/json/collection/v2.1.0/collection.json"
+  },
+  "item": [
+    {
+      "name": "Send new GPS coordinates to receiver",
+      "request": {
+        "method": "POST",
+        "header": [],
+        "body": {
+          "mode": "raw",
+          "raw": "{\r\n    \"trackedDeviceId\": \"ABBA\",\r\n    \"deviceType\": \"CLIENT_DEVICE\",\r\n    \"latitude\": 38.897675,\r\n    \"longitude\": -73.9729691,\r\n    \"speed\": 100\r\n}",
+          "options": {
+            "raw": {
+              "language": "json"
+            }
+          }
+        },
+        "url": {
+          "raw": "192.168.1.144:8082/gps-receiver/",
+          "host": [
+            "192",
+            "168",
+            "1",
+            "144"
+          ],
+          "port": "8082",
+          "path": [
+            "gps-receiver",
+            ""
+          ]
+        }
+      },
+      "response": []
+    }
+  ]
 }
