+++ ride-planner-service/src/main/java/ride/sharing/app/rideplannerservice/eventsender/KafkaEventSender.java
@@ -1,8 +1,8 @@
 package ride.sharing.app.rideplannerservice.eventsender;
 
-import com.ridesharing.domain.model.ride.RideDto;
-import com.ridesharing.domain.model.ride.events.EventType;
-import com.ridesharing.domain.model.ride.events.RideEvent;
+import com.ridesharing.rideplannerservice.RideDto;
+import com.ridesharing.rideplannerservice.events.EventType;
+import com.ridesharing.rideplannerservice.events.RideEvent;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.annotation.Profile;
