+++ ride-planner-service/src/test/postman/Ride Planning Service.postman_collection.json
@@ -1,68 +1,68 @@
 {
-	"info": {
-		"_postman_id": "ac770039-7f48-4e04-a4d1-46c860565061",
-		"name": "Ride Planning Service",
-		"schema": "https://schema.getpostman.com/json/collection/v2.1.0/collection.json"
-	},
-	"item": [
-		{
-			"name": "Get all rides",
-			"request": {
-				"method": "GET",
-				"header": [],
-				"url": null
-			},
-			"response": []
-		},
-		{
-			"name": "Create a request",
-			"request": {
-				"method": "POST",
-				"header": [],
-				"body": {
-					"mode": "raw",
-					"raw": "",
-					"options": {
-						"raw": {
-							"language": "javascript"
-						}
-					}
-				},
-				"url": {
-					"raw": "192.168.1.144:8082/rides",
-					"host": [
-						"192",
-						"168",
-						"1",
-						"144"
-					],
-					"port": "8082",
-					"path": [
-						"rides"
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "Update ride",
-			"request": {
-				"method": "PUT",
-				"header": [],
-				"body": {
-					"mode": "raw",
-					"raw": "{\r\n    \"pickupLocation\": \"Ankunft\",\r\n    \"destination\": \"Reiseziel\"\r\n}",
-					"options": {
-						"raw": {
-							"language": "json"
-						}
-					}
-				},
-				"url": {
-					"raw": ""
-				}
-			},
-			"response": []
-		}
-	]
+  "info": {
+    "_postman_id": "ac770039-7f48-4e04-a4d1-46c860565061",
+    "name": "Ride Planning Service",
+    "schema": "https://schema.getpostman.com/json/collection/v2.1.0/collection.json"
+  },
+  "item": [
+    {
+      "name": "Get all rides",
+      "request": {
+        "method": "GET",
+        "header": [],
+        "url": null
+      },
+      "response": []
+    },
+    {
+      "name": "Create a request",
+      "request": {
+        "method": "POST",
+        "header": [],
+        "body": {
+          "mode": "raw",
+          "raw": "",
+          "options": {
+            "raw": {
+              "language": "javascript"
+            }
+          }
+        },
+        "url": {
+          "raw": "192.168.1.144:8082/rides",
+          "host": [
+            "192",
+            "168",
+            "1",
+            "144"
+          ],
+          "port": "8082",
+          "path": [
+            "rides"
+          ]
+        }
+      },
+      "response": []
+    },
+    {
+      "name": "Update ride",
+      "request": {
+        "method": "PUT",
+        "header": [],
+        "body": {
+          "mode": "raw",
+          "raw": "{\r\n    \"pickupLocation\": \"Ankunft\",\r\n    \"destination\": \"Reiseziel\"\r\n}",
+          "options": {
+            "raw": {
+              "language": "json"
+            }
+          }
+        },
+        "url": {
+          "raw": ""
+        }
+      },
+      "response": []
+    }
+  ]
 }
